+++ src/svcs/flask.py
@@ -68,6 +68,7 @@ def register_factory(
     svc_type: type,
     factory: Callable,
     *,
+    enter: bool = True,
     ping: Callable | None = None,
     on_registry_close: Callable | None = None,
 ) -> None:
@@ -76,7 +77,11 @@ def register_factory(
     *app* that has been put there by :func:`init_app()`.
     """
     app.config[_KEY_REGISTRY].register_factory(
-        svc_type, factory, ping=ping, on_registry_close=on_registry_close
+        svc_type,
+        factory,
+        enter=enter,
+        ping=ping,
+        on_registry_close=on_registry_close,
     )
 
 
@@ -85,6 +90,7 @@ def register_value(
     svc_type: type,
     value: object,
     *,
+    enter: bool = True,
     ping: Callable | None = None,
     on_registry_close: Callable | None = None,
 ) -> None:
@@ -93,7 +99,11 @@ def register_value(
     that has been put there by :func:`init_app()`.
     """
     app.config[_KEY_REGISTRY].register_value(
-        svc_type, value, ping=ping, on_registry_close=on_registry_close
+        svc_type,
+        value,
+        enter=enter,
+        ping=ping,
+        on_registry_close=on_registry_close,
     )
 
 
@@ -101,6 +111,7 @@ def overwrite_factory(
     svc_type: type,
     factory: Callable,
     *,
+    enter: bool = True,
     ping: Callable | None = None,
     on_registry_close: Callable | None = None,
 ) -> None:
@@ -117,7 +128,11 @@ def overwrite_factory(
     """
     container = svcs_from()
     container.registry.register_factory(
-        svc_type, factory, ping=ping, on_registry_close=on_registry_close
+        svc_type,
+        factory,
+        enter=enter,
+        ping=ping,
+        on_registry_close=on_registry_close,
     )
     container.close()
 
@@ -126,6 +141,7 @@ def overwrite_value(
     svc_type: type,
     value: object,
     *,
+    enter: bool = True,
     ping: Callable | None = None,
     on_registry_close: Callable | None = None,
 ) -> None:
@@ -142,7 +158,11 @@ def overwrite_value(
     """
     container = svcs_from()
     container.registry.register_value(
-        svc_type, value, ping=ping, on_registry_close=on_registry_close
+        svc_type,
+        value,
+        enter=enter,
+        ping=ping,
+        on_registry_close=on_registry_close,
     )
     container.close()