+++ src/svcs/pyramid.py
@@ -105,6 +105,7 @@ def register_factory(
     svc_type: type,
     factory: Callable,
     *,
+    enter: bool = True,
     ping: Callable | None = None,
     on_registry_close: Callable | None = None,
 ) -> None:
@@ -113,7 +114,11 @@ def register_factory(
     *config*.
     """
     config.registry[_KEY_REGISTRY].register_factory(
-        svc_type, factory, ping=ping, on_registry_close=on_registry_close
+        svc_type,
+        factory,
+        enter=enter,
+        ping=ping,
+        on_registry_close=on_registry_close,
     )
 
 
@@ -122,6 +127,7 @@ def register_value(
     svc_type: type,
     value: object,
     *,
+    enter: bool = True,
     ping: Callable | None = None,
     on_registry_close: Callable | None = None,
 ) -> None:
@@ -130,7 +136,11 @@ def register_value(
     *config*.
     """
     config.registry[_KEY_REGISTRY].register_value(
-        svc_type, value, ping=ping, on_registry_close=on_registry_close
+        svc_type,
+        value,
+        enter=enter,
+        ping=ping,
+        on_registry_close=on_registry_close,
     )